Reorder files for `GetMessages` API results
