+++ app/src/main/java/com/infomaniak/mail/data/api/ApiRepository.kt
@@ -27,6 +27,9 @@ import com.infomaniak.mail.data.models.addressBook.AddressBooksResult
 import com.infomaniak.mail.data.models.correspondent.Contact
 import com.infomaniak.mail.data.models.draft.Draft
 import com.infomaniak.mail.data.models.draft.SaveDraftResult
+import com.infomaniak.mail.data.models.getMessages.GetMessagesByUidsResult
+import com.infomaniak.mail.data.models.getMessages.GetMessagesUidsDeltaResult
+import com.infomaniak.mail.data.models.getMessages.GetMessagesUidsResult
 import com.infomaniak.mail.data.models.message.DeleteMessageResult
 import com.infomaniak.mail.data.models.message.Message
 import com.infomaniak.mail.data.models.signature.SignaturesResult
